fixed logging bugs, load loggin conf from dict instead of file

+++ src/logger.py
@@ -9,12 +9,14 @@ import logging
 import logging.config,configparser
 
 
-def config(name=None,conf=None,**kwargs):
+def config(name=None,conf=None,file=None,write=None,**kwargs):
 	'''
 	Configure logging
 	Args:
 		name (str): File name for logger
-		conf (str): Path for logging config
+		conf (str,dict): Path or dictionary for logging config
+		file (str): Path to log file		
+		write (str):  Write type of conf file
 		kwargs (dict): Additional keywork arguments to overwrite config
 	Returns:
 		logger (logger): Configured logger
@@ -24,45 +26,59 @@ def config(name=None,conf=None,**kwargs):
 
 	logger = logging.getLogger(name)
 
-	default = 'log/logging.conf'
+	default = 'logging.conf'
 	conf = default if conf is None else conf
-	existing = os.path.exists(conf) if conf is not None else None
+	existing = os.path.exists(conf) if isinstance(conf,str) else None
 
-	file = kwargs.get('file')
 	path = os.path.abspath(os.path.expandvars(os.path.expanduser(file))) if file is not None else None
+	timestamp = datetime.datetime.now().strftime('%d.%M.%Y.%H.%M.%S.%f')
 	delim = '.'
-	ext = '%s.tmp'%(datetime.datetime.now().strftime('%d.%M.%Y.%H.%M.%S.%f'))
+	ext = '%s.tmp'%(timestamp) if write else None
 
 	props = ['loggers','handlers','formatters','keys']
-	keys  = {'class':'FileHandler'}
+	keys  = {'class':'logging.FileHandler'}
+	args = {'stream':{'sys.stdout':'ext://sys.stdout','sys.stderr':'ext://sys.stderr'}}
+	options = dict(disable_existing_loggers=False)
+	defaults = {'__name__':timestamp}
 	separator = '_'
 
 	if not existing:
-		source = os.path.join(os.path.abspath(os.path.dirname(__file__)),os.path.basename(default))
-		destination = os.path.join(os.path.abspath(os.path.dirname(conf)) if conf is not None else os.getcwd(),delim.join([os.path.basename(default),ext]))
-		directory = os.path.abspath(os.path.dirname(destination))
-		if directory not in ['',None] and not os.path.exists(directory):
-			os.makedirs(directory)
-		shutil.copy2(source,destination)
-		conf = destination
+		if write:
+			source = os.path.join(os.path.abspath(os.path.dirname(__file__)),os.path.basename(default))
+			destination = os.path.join(os.path.abspath(os.path.dirname(conf)) if conf is not None else os.getcwd(),delim.join([os.path.basename(default),ext]))
+			directory = os.path.abspath(os.path.dirname(destination))
+			if directory not in ['',None] and not os.path.exists(directory):
+				os.makedirs(directory)
+			shutil.copy2(source,destination)
+			conf = destination
+		else:
+			conf = os.path.join(os.path.abspath(os.path.dirname(__file__)),os.path.basename(default))
+	elif isinstance(conf,str):
+		if write:
+			source = conf
+			destination = delim.join([os.path.basename(default),ext])
+			directory = os.path.abspath(os.path.dirname(destination))
+			if directory not in ['',None] and not os.path.exists(directory):
+				os.makedirs(directory)
+			shutil.copy2(source,destination)
+			conf = destination
+		else:
+			conf = str(conf)
 	else:
-		source = conf
-		destination = delim.join([os.path.basename(default),ext])
-		directory = os.path.abspath(os.path.dirname(destination))
-		if directory not in ['',None] and not os.path.exists(directory):
-			os.makedirs(directory)
-		shutil.copy2(source,destination)
-		conf = destination
+		conf = dict(conf)
 
 	if conf is not None:
 		try:
-			config = configparser.ConfigParser()
-			config.read(conf)
+			if isinstance(conf,str):
+				config = configparser.ConfigParser()
+				config.read(conf)
+			else:
+				config = logging.config.dictConfig(conf)
 
 			sections = [section for section in config if all(config[section].get(key) == keys[key] for key in keys)]
 			for section in sections:
 
-				name = section.split(separator)[-1]
+				string = section.split(separator)[-1]
 
 				if file is None:
 
@@ -71,21 +87,20 @@ def config(name=None,conf=None,**kwargs):
 						for prop in props:
 							if prop not in config[section]:
 								continue
-							if name in config[section][prop]:
-								names = config[section][prop].split(',')
+							if string in config[section][prop]:
+								strings = config[section][prop].split(',')
 								config[section][prop] = ','.join([
-									i for i in names
-									if i not in [name]])
+									i for i in strings
+									if i not in [string]])
 				
 					sections = list(config)				
 					for section in sections:
 						subname = section.split(separator)[-1]
-						if (subname == name):
+						if (subname == string):
 							config.pop(section);
 							continue
 					
 				else:
-
 					directory = os.path.dirname(path) if path is not None else None
 					if (directory is not None) and (not os.path.exists(directory)):
 						os.makedirs(directory)
@@ -93,9 +108,9 @@ def config(name=None,conf=None,**kwargs):
 					prop = 'args'
 					config[section][prop] = '(%s)'%(','.join([
 						'"%s"'%(path),
-						# '"%s"'%('"a"'),
 						*(config[section][prop][1:-1].split(',')[1:] 
-						if not os.path.exists(path) else ['"a"'])]
+						if not os.path.exists(path) else ['"a"'])
+						]
 						))
 
 			sections = list(config)				
@@ -108,19 +123,17 @@ def config(name=None,conf=None,**kwargs):
 						config.pop(section);
 						break
 
-
 			sections = list(config)				
 			for section in sections:						
 				for prop in props:
 					if prop not in config[section]:
 						continue
-
-					names = config[section][prop].split(',')
-					for name in names:
-						if not any(name in section.split(separator)[-1] for section in config if separator in section):
+					strings = config[section][prop].split(',')
+					for string in strings:
+						if not any(string in section.split(separator)[-1] for section in config if separator in section):
 							config[section][prop] = ','.join([
-								i for i in names
-								if i not in [name]])
+								i for i in strings
+								if i not in [string]])
 					if not config[section][prop]:
 						config.pop(section);
 						break
@@ -135,48 +148,99 @@ def config(name=None,conf=None,**kwargs):
 						config.pop(section);
 						break
 
-			with open(conf, 'w') as configfile:
-				config.write(configfile)
-			
-			try:
-				logging.config.fileConfig(conf,disable_existing_loggers=False,defaults={'__name__':datetime.datetime.now().strftime('%d.%M.%Y.%H.%M.%S.%f')}) 	
-			except KeyError:
-				pass
+
+			if write:
+
+				with open(conf, 'w') as configfile:
+					config.write(configfile)
+				
+				options.update(dict())
+
+				try:
+					logging.config.fileConfig(conf,**options) 	
+				except KeyError:
+					pass
+
+			else:
+
+				conf = {}
+				sections = set(section for section in ['%ss'%(section.split(separator)[0]) if separator in section else section for section in config.sections()] if section in props)
+				strings = {'root':''}
+				for section in sections:
+					conf[section] = {}
+					for string in config.sections():
+						attr = string.split(separator)[-1]
+						item = '%ss'%(string.split(separator)[0])
+						if separator not in string or item != section:
+							continue
+						attr = strings.get(attr,attr)
+						conf[section][attr] = {}
+						for key, val in config.items(string):
+
+							if section in ['handlers'] and key in ['args']:
+								key,val = None,[i.strip() for i in val.replace('(','').replace(')','').replace('"','').split(',') if i]
+								for arg in args:
+									if all(v in args[arg] for v in val):
+										key,val = [arg],[args[arg].get(v,v) for v in val]
+										break
+								if key is None:
+									key,val = ['filename','mode'],[*val]
+							else:
+								key = [key]
+								val = [val]
+							for k,v in zip(key,val):
+								conf[section][attr][k] = v
+
+				for section in conf:
+					for attr in conf[section]:
+						for key in conf[section][attr]:
+							if section in ['loggers'] and key in ['handlers']:
+								conf[section][attr][key] = conf[section][attr][key].split(',') if isinstance(conf[section][attr][key],str) else [*conf[section][attr][key]] 
+						if any(f'%({default})' in conf[section][attr][key] for default in defaults for key in conf[section][attr] if isinstance(conf[section][attr][key],str)):
+							conf[section][attr]['defaults'] = defaults
+
+				options.update(dict(version=1))
+				conf.update(options)
+
+				try:
+					logging.config.dictConfig(conf) 	
+				except KeyError:
+					pass
 
 		except Exception as exception:
+			print(exception)
 			pass
 
 		logger = logging.getLogger(name)
 
-
-	try:
+	if write:
 		try:
 			os.remove(conf)
 		except:
 			pass
-		# if (conf is not None) and ((file is None) or (os.path.dirname(conf) != os.path.dirname(file))) and (os.path.exists(os.path.dirname(conf))) and (not os.listdir(os.path.dirname(conf))):
-		# 	try:
-		# 		os.rmdir(os.path.dirname(conf))
-		# 	except:
-		# 		pass
-	except:
-		pass
+		if (conf is not None) and ((file is None) or (os.path.dirname(conf) != os.path.dirname(file))) and (os.path.exists(os.path.dirname(conf))) and (not os.listdir(os.path.dirname(conf))):
+			try:
+				os.rmdir(os.path.dirname(conf))
+			except:
+				pass
 
 	return logger
 
 
 class Logger(object):
-	def __init__(self,name=None,conf=None,file=None,cleanup=None,verbose=None,**kwargs):
+	def __init__(self,name=None,conf=None,file=None,write=None,cleanup=None,verbose=None,**kwargs):
 		'''
 		Logger class
 		Args:
 			name (str,logger): Name of logger or Python logging logger
-			conf (str): Path to configuration
+			conf (str,dict): Path or dictionary to configuration
 			file (str): Path to log file
+			write (str):  Write type of conf file
 			cleanup (bool): Cleanup log files upon exit
 			verbose (int,str,bool): Verbosity
 			kwargs (dict): Additional keyword arguments
 		'''
+
 		if (name is None) or isinstance(name,str):
 			try:
 				self.logger = config(name=name,conf=conf,file=file,**kwargs)
@@ -254,7 +318,7 @@ class Logger(object):
 		for logger in loggers:
 			try:
 				os.remove(logger)
-				# os.rmdir(os.path.dirname(logger))
+				os.rmdir(os.path.dirname(logger))
 			except:
 				pass
